backend/backend_config/wsgi.py
from dj_static import Cling

import os

from django.core.wsgi import get_wsgi_application

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'backend_config.settings')

# application = get_wsgi_application()
application = Cling(get_wsgi_application())