backend/prod-requirements.txt
africastalking==1.2.9
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
django==4.2.21
djangorestframework==3.15.1
django-cors-headers==4.3.1
idna==3.6
Pillow==10.2.0
python-decouple==3.8
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.1